move images, update imports
