+++ src/components/Header/index.tsx
@@ -2,8 +2,8 @@ import { useState } from "react";
 import Button from '@material-ui/core/Button';
 import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
-import SunflowerLogo from "../../img/sunflowerlogo.png";
-import { ReactComponent as Hamburger } from "../assets/hamburger.svg";
+import SunflowerLogo from "../../assets/sunflowerlogo.png";
+import { ReactComponent as Hamburger } from "./assets/hamburger.svg";
 import "./Header.css";
 
 export const Header = () => {
