+++ src/pages/MyGardens/index.tsx
@@ -1,4 +1,4 @@
-import { Header } from "../../components/Header/Header";
+import { Header } from "../../components/Header";
 
 export const MyGardens = ()  => {
   return (
